% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twophase.R
\name{evalTwoPhase}
\alias{evalTwoPhase}
\title{evalTwoPhase functions}
\usage{
evalTwoPhase(Outcome, X, Z, Stratum, Phase_ID, namesX, namesZ, q, p, method)
}
\value{
3 numbers
}
\description{
evalTwoPhase functions
}
\examples{

library(package1014)
library(mvtnorm)
set.seed(2016311033)













}
