% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twophase.R
\name{seTwoPhase}
\alias{seTwoPhase}
\title{seTwoPhase functions}
\usage{
seTwoPhase(
  Outcome,
  X,
  Z,
  Stratum,
  Phase_ID,
  namesX,
  namesZ,
  q,
  p,
  numBoot,
  method
)
}
\value{
3 numbers
}
\description{
seTwoPhase functions
}
