% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twophase.R
\name{summaryTwoPhase}
\alias{summaryTwoPhase}
\title{summaryTwoPhase functions}
\usage{
summaryTwoPhase(eval_vec, se_vec, q, p, method)
}
\value{
return the summary. TPR,FPR,AUC.
}
\description{
summaryTwoPhase functions
}
